добавлены права доступа, генератор кодов, ui
реализованы права доступа для удаление и изменения книг, глав, томов и пользователей
реализован генератор кодов для восстановления пароля
добавлен временный консольный интерфейс

+++ src/main/resources/sql/schema.sql
@@ -2,7 +2,7 @@ CREATE DOMAIN email AS text
 CHECK (VALUE ~* '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+[.][a-zA-Z]{2,}$');
 
 CREATE TABLE IF NOT EXISTS users(
-id serial PRIMARY KEY,
+id bigserial PRIMARY KEY,
 nickname VARCHAR(30) NOT NULL UNIQUE,
 email email NOT NULL UNIQUE,
 password_hash VARCHAR(255) NOT NULL,
@@ -10,7 +10,7 @@ created_at TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS books(
-id serial PRIMARY KEY,
+id bigserial PRIMARY KEY,
 title text NOT NULL,
 date_changed TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP,
 created_at TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
@@ -35,25 +35,25 @@ volume_id INTEGER REFERENCES volumes(id) ON DELETE CASCADE NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS books_authors(
-user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-book_id INTEGER REFERENCES books(id) ON DELETE CASCADE,
+user_id bigint REFERENCES users(id) ON DELETE CASCADE,
+book_id bigint REFERENCES books(id) ON DELETE CASCADE,
 PRIMARY KEY(user_id, book_id)
 );
 
 CREATE TABLE IF NOT EXISTS genres(
-    id serial PRIMARY KEY,
+    id bigserial PRIMARY KEY,
     name VARCHAR(100) UNIQUE NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS books_genres(
-book_id INTEGER REFERENCES books(id) ON DELETE CASCADE,
-genre_id INTEGER REFERENCES genres(id) ON DELETE CASCADE,
+book_id bigint REFERENCES books(id) ON DELETE CASCADE,
+genre_id bigint REFERENCES genres(id) ON DELETE CASCADE,
 PRIMARY KEY(book_id, genre_id)
 );
 
 CREATE TABLE IF NOT EXISTS password_reset_codes(
 id serial PRIMARY KEY,
-user_id INTEGER REFERENCES users(id) ON DELETE CASCADE NOT NULL,
+user_id bigint REFERENCES users(id) ON DELETE CASCADE NOT NULL,
 code VARCHAR(10) NOT NULL,
 expires_at TIMESTAMPTZ NOT NULL,
 created_at TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
